requirements.txt
allure-pytest==2.8.16
allure-python-commons==2.8.16
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
idna==2.10
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-rerunfailures==9.0
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
urllib3==1.25.9
wcwidth==0.2.5
xlrd==1.2.0
xlutils==2.0.0
jsonpath_rw==1.4.0
openpyxl==3.0.6